SPARK does not allow anonymous types while Ada allows them in many places:
for both the type of index and the type of element in an array type; directly in a
definition of a variable (for array or scalar type); as the type of a component
in a record definition. This extension consists in translating uses of anonymous
types in Ada in uses of named types in SPARK, with a preceding definition for
the new named types created.