Remove useless why function for 'Mod
As the frontend expands 'Mod into a regular mod operation, there is
little point in having a Why function that we (probably) won't have to
emit in any case. Remove it and document this removal. Add a test that
illustrate this; and take this occasion to reformat the test output... in
order to make some sense out of it.

Part of J526-010.