Logical operators for modular types
Add an incomplete description of how or/and/xor for modular types
would be represented in Why. Incomplete, because we do not have a
correct axiom system for bitwise operations for now; this could be
done using conversions to one-dimensional boolean arrays, but these
are not described yet.

Part of J526-010.