Float/fixed conversions
After having add a closer a look at the expanded tree for 'Round, it
seems that a "user-level" float/fixed conversion would not be very
useful; the expander already does some of the job. What we really need
is a int/real cast with the appropriate rounding policy.

Part of J526-010.