Range invariant at update
The mathematical model of arrays relocates all arrays to 0 .. Length - 1
and (therefore) offers no way to express "strict" range equalities.
In other words, two Ada arrays may be converted to the same logic
array even though they have a different range, so when converting back
and forth we must check the range is equal for a "strict" equality.
This "strict" equality should in particular be enforced when updating
an element in an array; add the corresponding invariant for this case.
That also means that the unicity axiom does not hold.

Part of J526-010.