Rewrite predicates as logic functions
Replace predicates by logic functions that return bool. Two reasons
for that: first, logic functions have a better behavior in terms of
proof (one more trigger for equality) and allow to use equality
between abstract type more extensively; second, logic functions can be
used in the program space whereas predicate cannot --and we have no
particular reason to restrict these functions to the logic space.
This change was meant to evaluate the impact of such a scheme:
the most visible effect is the simplification of pre/post on array types.

This choice was discussed on the why-discuss mailing list:
http://lists.gforge.inria.fr/pipermail/why-discuss/2010-July/000507.html

Part of J526-010.