Multi-dimensional arrays
Add a model for multi-dimensional arrays. So far it defeats some
provers (and particularly alt-ergo); it should be improved later.
Whether that improvement should go in alt-ergo or in this model is
still to discuss.

The model is illustrated by a version of the linear search example for a
two-dimension table.

Part of J526-010.