Unicity axiom and equality in programs
The previous revert has made the unicity axiom inoperant; re-implement
to help alt-ergo to use it, i.e. without using a predicate. That
is a way to retrieve the first advantage of logic functions. The second
advantage was that a logic function could be used in program; in order to
get that back with predicates, just define a new parameter eq_<type>_
that uses the eq predicate in the postcondition.

Part of J526-010.