Rename some fields to avoid conflict with keywords
The builders for Why node will be automatically generated, and the
param names in the builder functions will be the field name in the
variant part with their prefix stripped (e.g. 'Name' for 'RS_Name'
when building a node whose kind is W_Raise_Statement). To facilitate
the work of the automatic generation, rename the field name that can
conflict with Ada keywords after having stripped the prefix.

Part of J719-024.