Support for templates of Ada packages
This adds some support for reading a template of an Ada package and
replace some patterns by some generated code. It is used by xtree and
xkinds; it replaces the -head/-tail scheme by a more generic scheme
that we will help when we will make them generate some other package
specs and bodies. No functional change so far, though.

Part of J719-024.