Generate declarations of Option Id types
To mark that a child of a Why node would be optional, we used to
initialize it to Why_Empty or Why_Empty_List. It is still to be
discussed if default values in record declarations will be supported by
hi-lite; so use an other caracterisation. Have an Id subtype whose
predicate will ensure that its values will either be of a given kind,
or empty. These will be named Option Id types (or Option List types).
Note that this will also allow to discriminate between both cases (can
be empty or not) for parameters of builder subprograms.

Part of J719-024