Clean up empty node/list handling
In order to avoid to have a valid node at index Empty in the node
table, create an invalid entry at this index, at initialization
time. For node lists, do not have a preallocated singleton index for
empty lists; it is much simplier to allocate new ids in the list table
for each empty list to simplify append operations.

Part of J719-024