Generation of recursive traversal
This adds the generation of recursive traversal through a Why syntax
tree.  It is directly inspired by the model provided by ASIS.Iterator,
with the difference that it uses tagged types and dispatching instead
of access to procedures; tagged types are more "hi-lite-friendly" and
their use would make it easier to bootstrap the backend (some parts,
at least) some day. An other (temporary) difference with ASIS.Iterator
is that traversal control has not been implemented (yet); it should
follow in forthcoming changes.

Part of J719-024