Kind-validity
Generate a set of parameterized expression that allows to check that
a given id is of a given kind, with the appropriate multiplicity
(node, non-empty node, list, non-empty list). These will be used
as subtype predicate to the corresponding subtypes as soon as
subtypes predicates are supported by GNAT.

Part of J719-024.