List validity
Implements checks for node lists, using a cache that will be updated
by the operations that modifies the list (e.g. append). This has the
advantage to be more efficient than a full traversal of the list
each time we check validity; and has the other advantage to be
implementable now, whereas universal quantif is not supported by
GNAT yet.

Take this opportunity to fix the link to parent for list by
introducing a new type, Why_Node_Set, that is the father of both
Why_Node_Id and Why_Node_List. This link will be useful if we
implement an upward recursive update of a validity from a given node;
it was not usable for lists before this change (even though its
documentation was saying that it was).

Part of J719-024