Part of J908-013
This extension consists in fixing the missing case of anonymous extension.
The transformation of discrete range (Transform_A_Discrete_Range)
generates automatically the subtype mark (i.e Integer) of the index in
the following case :
     Ten_Characters : String(1 .. 10);
     for I in 1 .. 10 loop;
This tranformation don't work in the following example where the
subtype mark is Stack_Range and not Integer.
     type Stack_Range is range 0 .. Stack_Size;
     type Vector is array(Stack_Range range <>) of Integer;
     Stack_Vector : Vector (1 .. Stack_Size);

* sparkify/sparkify-post_operations.ad(s|b)

(Transform_A_Discrete_Range) : update the function  with new optional
  argument (Subtype_Decl, initialized with Nil_Element) which specified
  if the subtype of discrete range is known.

(Transform_An_Index_Constraint): call the previous function with the
  right declaration of the subtype mark if already presents.

Update Print_All_Use_Type
to print Use_Type clause of types (visible only in a data package)
in internal and external package.