Simple validity policy for lists of Why AST nodes
We only check that these are not empty when they shall not; the
kind-validity of each element is assured by preconditions in
append/prepend defined in why-atree-mutators and never checked after
that; if this API is used, then the list is valid. This should be a
good compromise for now; we'll see if we want to do something more
elaborated when we have universal quantifiers on containers.

Part of J719-024