Why AST - fix a thinko in syntax spec
The syntax, as it was specified by Why.Atree, was only accepting
logic function declarations as valid declarations in logic space. For
example, abstract type declarations were illegal... This bug was
detected while adding some additional checks to builders/mutators; it
is fixed by the present commit.

Part of J719-024