[Why AST] Array of Ids in builders
As stated in the previous commit, use array of ids to represent lists
in builder parameters. This should allow us some simplifications
in various places in why-gen-*. For now, the pre/post conditions
do not check for the array validity; but this is checked in the
builder body by an assertion.

Part of J719-024