* alfa/alfa-common.ads New package for common utilities across ALFA packages, like sets of entities or helper functions
* alfa/alfa-definition.adb, alfa/alfa-definition.ads
New package to flag entities inside/outside ALFA

* alfa/alfa-filter.adb
(Bucket_Dispatch): remove initializing expression of global constant if not in
ALFA

* alfa/alfa-frame_conditions.adb, alfa/alfa-frame_conditions.ads
cleanup

* why/gnat2why-driver.adb
(GNAT_To_Why): call ALFA.Definition subprograms to mark entities outside ALFA
(Translate_List_Of_Decls): add comments
(Translate_Standard_Package): add translation of entities from Standard which
  are internal

* why/gnat2why-subprograms.adb
renaming

For K302-028

A pass through all units is performed to detect which entities fall outside
the ALFA boundaries. Units should be seen in an order that guarantees no
forward references, so that an entity can be flagged as not in ALFA when seeing
its declaration, before any use of this entity is flagged as not in ALFA.

The various reasons for being out of ALFA are monitored: either because the
current implementation does not allow it, or because of a specific restriction
which could be raised in the future, or else because of the main design of
ALFA. These reasons are propagated, so that an occurrence of an object not in
ALFA in a subprogram leads to the subprogram body not being in ALFA for the
same reasons as the object. This will allow early feedback from users on
the most important extensions of ALFA.