* alfa/alfa-definition.adb (Formal_Proof_Currently_Forced): detect whether a scope in the current stack   has formal proof forced, with no inner scope having formal proof disabled (Current_Scope): refactor code for scopes. Now packages, subprograms and   objects are all scopes. Loops are stored in the scope stack but skipped by   Current_Scope.
* why/gnat2why-driver.adb
(GNAT_To_Why): remove special handling of errors

Refactor code so that pragma Annotate (Formal_Proof, On) can be put anywhere,
and it determines whether the current package or subprogram has formal proof
forced or disabled. Such pragmas can be reversed in an inner scope. Useful to
understand why some entity is not in ALFA, as you get the error message if
you set formal proof to On.