* alfa-filter.adb (Bucket_Dispatch): create a subprogram declaration if not already present, in   order to generate a corresponding forward declaration (Filter_Compilation_Unit): sort subprogram declarations for the body so that   subprogram declarations precede subprogram bodies
* gnat2why-decls.adb
(Full_Name): move essential of function to Sem_Util.Unique_Name

* gnat2why-driver.adb
(GNAT_To_Why): remove late qualification of names, now done in frontend

* gnat2why-subprograms.adb
Use Full_Name of identifier everywhere. Remove special case for body serving as
declaration, which was leading to late declaration of subprogram.

* why-gen-progs.adb, why-gen-progs.ads
(New_For_Loop): change signature to take an identifier instead of a name for
  the loop index

For K323-029

Move Full_Name to Sem_Util.Unique_Name, so that the frontend knows the final
name of objects when creating cross-references in the ALI file. Correct
the generation of forward declarations for subprogram bodies serving as
declaration.