First implementation of Why.Gen.Binders
For now, this just re-defines builders provided by Why.Gen.Decl on a
common basis: a type Binder_Type that abstracts away binders for the
logic space and the program space. It uses this new API in a couple of
places just to make sure that it works. It also declares the
(not-yet-implemented) subprogram Emit_Top_Level_Declarations that is
meant to help generating series of top-level declarations from the
same Ada entity.

Part of K415-020