Default domains
For certain node kinds, the domain is known a priori and should be set
automatically (e.g. universal quantifs are always in the predicate
domain). This commit adds some support of default domains to nodes in
order to make that happen. For now, only the default value of Domain
in builders are changed; the goal is ultimately to remove this
parameter when its value is not ambiguous.

[K810-007]