Do not generate a ref binding for the quantified var of a quantified expression
We translated the quantified variable of a quantified expression using a ref
binding, to be coherent avec simple loop parameters. But as quantified
expressions can appear in predicates, this is incorrect. This just happens to
work with Why2, but doesn't with Why3. Now only loop parameters of real loops
are mutable.

For K211-002