Support for abstract types in base types
This is meant to replace the Why_Type structure and use a AST node
instead; W_Base_Type nodes are almost exactly what we need, except
that it does not handle abstract types. Add support for such types;
then, in such a case, the type name is deduced from the corresponding
Ada node.

For now, W_Abstract_Type nodes are not removed, as they are still a
bit tricky to eliminate; but the switch to Why 3 should simplify this
work by removing most of the code the generates general theories. For
now, just mark W_Abstract_Type as deprecated.