For K919-014
Frontend sometimes declares in package B an Itype for the base type of a type
T declared in package A. Thus we could get a conflict when two such packages
B1 and B2 were declaring the same type. Now, Itypes serving as base types are
always marked at the point of declaration in A, so B1/B2 do not declare it.