For K303-009
Deal with all kinds of record aggregates, including with others and box
components. Others are simply replaced by appropriate associations by
front-end, but box are either removed from aggregate (in program) or left as
box association (in contracts). Currently deal with these cases.

Equality between records is translated into primitive equality in Why, with
an axiom stating this is the same as component-wise equality.