Support for complex expressions together with attribute 'Old
Previously, only variables could be used with 'Old; now, we allow any
expression.

This is easy with Why3, which supports the construct
  (old <expr>)
for complex expressions.

However, the current Syntax tree had to be adapted to allow "tags" (formerly
called labels, but this clashes with text labels) to be applied to expressions
and not only identifiers

For K506-012