Generate checks for postconditions
We do so by generating an ignored program expression in the return exception
handler. The main problem is the translation of 'Old expressions; we translate
such expressions by an identifier, and store this ident together with the
corresponding expression in a list; this list is then used to generate
bindings for these expressions at the beginning of the subprogram.

Also, remove some obsolete comments, and rename Why_Decl_Of_Ada_Subprogram to
Transform_Subprogram.

For K420-006