Introduce class W_Value
To discriminate between things that are either progs or terms but really are
in program space, from things that may exist in both program and logic space
and would accept prog on side and terms on the other. The first one would be
of kind W_Prog, whereas the second one would be of kind W_Value.

Part of K923-008