Declare a logic function for all Ada functions, not only expression functions. An axiom is still generated only for expression functions.
Take into account global reads as additional arguments for the logic function.

This requires adding a parameter Ref_Allowed to Transform_Expr and all similar
functions which may be called both in the context of a program, and in the
context of a pure logical declaration (axiom, logic function). In the first
case, references are allowed. In the second case, all references should be
replaced by parameters of logic functions and quantified variables in axioms.