Use abstract types for subtypes of Standard.Boolean
For any subtype of an enumerated type, we have to use an abstract type
in order to have range checks at conversion time. In the case of Boolean,
it still makes sense, since one may have a (quite useless) boolean type
ranging from True to True. We use the type bool directly in the case
of bool.

This does not bring new regressions, so it is OK for now; but one may
note that the model for Boolean subtypes is the same as the one for
any enumeration type, relying on int; it would make sense to improve
that by having a dedicated model relying on bool. Plus, one may note
that any non-pathological boolean type (i.e. containing both True and
False) could probably be modeled as bool directly. Whatever we choose,
we should probably wait for Why 3 theory to work before moving forward
on these optimizations.

[KA20-063]