Update test and output. Hosting of expressions from invariant is fine.
For K516-008

The recent disabling of the generation of checks in the frontend leads to no
expressions being hosted of invariants, even with calls. So TN is fixed. If a
more complex case leads to expressions being hosted, a new TN will be opened
(and this is safe, as it just leads to an invariant being considered like an
assertion).