Follow-up of KC14-026
The use of iterators (instead of a loop specification) in a quantified
expression is marked as not yet implemented in Alfa.

An axiom is generated for an expression function, to give it a postcondition
corresponding to its body, only when the body is in Alfa. This involves moving
a few functions to gnat2why-decls.