(mintapif): Add optional "prealloc" mode where the receive buffer is a fixed preallocated SPLIT_BUF for the maximum frame size, to simulate a DMA-based driver.
(mintapif:low_level_output, ne2kif:low_level_output): Fix incorrect condition
in output loop. For a SPLIT_BUF, the Len is equal to the Tlen, and covers
all data up to the end of the buffer, so exit even if not at end of the
buffer chain, as soon as that many bytes have been written.

For L207-005


git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/ipstack@185421 936e1b1b-40f2-da11-902a-00137254ae57