Accepting this regression
For KC14-026

This one is really a mistery. It actually gets proved, but needs longer than
the time we give in the testsuite (1s). Normally, with all the context
reduction we did, it should be *faster* to run, but maybe some of the removed
context helped alt-ergo find things faster ... SMT solver oddness +1!

It's not really a regression of the tool anyway, as with a realistic timeout
the VC is OK.