First prototype of a tool to merge test/proof results
In Python, in order to quickly validate the topos-based approach;
the idea of this approach is to be able to apply lattice-based
operations (meets, joins) on complicated structure that are not
know a priori. For now, we suppose that this structure can be
modelized by a partial order. i.e. in the case implemented here,
(PROVED, TESTED) > OK > SUBPROGRAM.

Checking this in only for the record; this does not seem to be the
appropriate approach. Although this solution presents an elegant way
to separate (1) the structure of the subobjects (described in the
sketch), (2) the actual contents and (3) the join/meet operations,
this is probably a bit too general for our problem of merging
test/proofs results.

Part of L222-014