Lattice-based prototype for merging test/proof results
Rewrite the prototype to drop topoi and use much simpler structures
instead. The idea of representing the status with a partial order
still holds, and we complete it to have a lattice. The set of sloc
ranges can also be seen as a lattice, so this should allow some code
sharing.

Part of L222-014