Treat private types as abstract types when knowledge of the underlying type is not needed. This allows formally analyzing client code of packages even if the actual type is not in Alfa. Note that some subprograms from the package defining the private type may also be in Alfa if they do not access the underlying non-Alfa type.
As part of this work, remove the distinction between unique entities and
entities. Many calls to Unique are also removed (but some are still necessary,
to be cleanup up later).

When some constructs require that the most underlying is in Alfa, mark it so
in Alfa.Definition by calling explicitly Mark_Most_Underlying_Type_In_Alfa.

No private type is ever translated to Why. If the most underlying type is in
Alfa, it is used in place of the private type. Otherwise, the special __private
type is used (for which a new base type EW_Private is defined). This also
requires treating record subtypes as if they were designating directly their
base type.

For KB29-038