In order to support deferred constants, and parameterized proofs of client units not knowing the value of the constant, we now define separate modules for the deferred constant and its completion. The module for the completion is marked as a completion of the module for the deferred constant, so that further modules including the base theory also include the completion, for modules defined in this unit. Theories defined in other units only have access to the base theory.
This also forced a change in Alfa.Definition: we consider entities for
variables instead of unique entities. This is in fact cleaner: all references
to an entity always refer to the same entity (here, the deferred constant, not
the completion), only the object declaration may refer to the completion. The
same seems true of other entities, to be cleaned up later.

For L320-006