* gnat_backends/alfa/alfa-definition.ad(b|s)( (All_Entities): Make the list public, so that translation to Why3 can use it. (Pop_Scope): For the special case of an expression function, use as entity for   the body the distinct E_Subprogram_Body entity.
* gnat_backends/why/gnat2why-decls.ad(b|s)
* gnat_backends/alfa/alfa-util.ad(b|s)
Move to alfa-util subprograms previously defined in gnat2why-decls.
Give a correct Domain argument to To_Why_Id instead of the default WE_Prog.

* gnat_backends/why/gnat2why-driver.adb
(Do_Generate_VCs): Separate procedure for generating VCs for subprograms.
(Translate_CUnit): Before translating entities for the unit, take into account
  the expression functions from withed units, by declaring completions for
  them, so that the defining axioms are available. Also, separate the
  translation of entities (which may define completions) from the generation of
  VCs, which uses these completions.
(Translate_Entity): Remove generation of VCs. Add a special translation for
  E_Subprogram_Body entities, which should correspond to expression function
  bodies.

* gnat_backends/why/gnat2why-nodes.ad(b|s)
(In_Some_Unit_Body): New query function.

* gnat_backends/why/gnat2why-subprograms.ad(b|s)
(Translate_Expression_Function_Body): Make procedure public. Make it create
  a new theory, which completes the previously existing one declaring the
  logic function.

* gnat_backends/why/why-inter.ad(b|s)
(Add_Completion, Get_Completions): Change signature, so that we distinguish
  completions in context-spec from completions in context-body.
(Close_Theory): Add completions from both context-spec and context-body.

For KC14-023

We need a separate definition of theories in Why3 for declaring a logic
function and its axiom. This is necessary so that they are defined with the
proper visibility over previously defined entities, and other entities have
visibility over the function. Both theories may be defined in the spec, or in
the body, or the decl in the spec and the def (axiom) in the body. To that
effect, an entity E_Subprogram_Body is added to the lists of entities to
translate in Alfa.Definition, which is then treated specially in
Translate_Entity. The theory defining the axiom is seen as a completion of the
logic function theory: after defining the axiom, all includes of the function
theory also include the axiom theory. For this to have an effect on VCs
generated, VC generation must be postponed until after completions have been
defined, hence the change in Translate_CUnit/Translate_Entity. Also,
completions for expression functions in other units must be taken into
account. Finally, completions may come either from context-spec or
context-body.