* alfa/alfa-definition.ad(b|s) (Mark_Type_Entity): Mark needed types before the current type.
Type declarations may refer to private types whose full view has not been
declared yet. However, it is this full view which may define the type in Why3,
if it happens to by in Alfa. Hence the need to define it now, so that it is
available for the current type definition. So we start with marking all needed
types if not already marked, before the current type.

This corrects the half dozen regressions seen since Friday on ACATS in mode
gnatprove.