Fix incorrect dependency computation
When tmp variables are introduced for "copy back" on subprogram calls, we
use the name of the formal to generate a name for the variable. However, we
should not include the *entity* of the formal, because this would imply a
dependency on that formal, which is not the case.

For K831-027