Change the generation of logic function/axiom for aggregates and slices
Previously, we generated a logic function that took as parameters all mutable
objects occurring in the aggregate or slice, and we called this function only
for translations as terms. For translations as progs, we used an "any"
construct instead, which was buggy because it skipped some checks. Another
problem with this translation was that it was mostly done in some fixed
context, so 'Old and 'Result were badly translated in some cases.

The new translation solves these problems by generating a logic function which
takes all needed sub-expressions (components for aggregates, prefix and bounds
for slices) as parameters. The actual translation of these sub-expressions only
occurs at the point of call, so in the proper context.

For L427-012

* gnat_backends/why/gnat2why-expr.adb
(Transform_Aggregate, Transform_Slice): New functions that replace the previous
  Transform_Array_Value_To_Term, so that the behavior is tailored for
  aggregate or slice (although these functions are close in terms of behavior).
  Also define locally some helper subprograms previously defined at the level
  of the enclosing package.

* gnat_backends/why/gnat2why-nodes.ads
Remove unneeded data structures