gnatmerge: add new order to lattices to properly represent goals
The model of a merge of several status is a lattice; but the partial
order of this lattice is different from the one that allows to compare
a status to a goal. Typically, in the first one, PARTIALLY OK will be
stronger than OK and KO, as it contains at least one results of both;
whereas the goal order shall obviously be KO < PARTIALLY OK < OK. To
avoid any confusion, rename the first order is_in; less_than will be
used for the second one.

[L809-016]