Only run gnatwhy3 on files that belong to current project
This is achieved by querying GNATCOLL.Projects interface for all the "library
files" with suffix "__package.mlw" whose base name is part of the project.
This list is then put in why.gpr, the file that is passed to gprbuild for the
last phase of gnatprove (calling gnatwhy3). Previously, why.gpr contained only
a Source_Dir attribute, not a Source_Files attribute.

gnatprove.adb:
   (Generate_Project_File) take a list of files instead of a directory as
   argument, generate Source_Files list instead of Source_Dir attribute;
   (Generate_Why_Project_File) compute list of "__package.mlw" files, and pass
   that list to Generate_Project_File;
   (Compute_VCs) Object_Dir is not needed anymore for
   the call to Generate_Why_Project_File, instead pass the project_type

For L097-021