LA08-002 use Ada object name instead of generic name
Previously, we generated names s1__s2__s3__x.obj for variable names in Why3,
where the s_i are the different scopes of the variable declaration, and x the
variable name. "obj" was a generic suffix. This is now changed to use the
variable name x instead of "obj" as the actual Why3 name for the object.

This does not change much the readability of Why3 code, but makes Why3 use
that variable name instead of "obj" as the basis for the VC name of the
object. Overall, this increases readibility of the VCs for debugging.

string_utils.ads/adb:
   (Lower_Case_First): procedure to lowercase the first char of a string, if
      not already the case
   (String_Sets): new instance of Ada container fro sets of strings

why-inter.adb:
   (Avoid_Why3_Keyword): new function to add "__" to a string if matches a
      why3 keyword. Is used to make accidental use of Why3 keywords as
      variable names impossible
   (Extract_Object_Name): Some variable names are known to gnat2why as
      fully expandad strings only (s1__s2__s3__x). This new function searches
      for the last "__" and returns the substring after it. This is considered
      the "short" object name for that object.
   (To_Ident) WNE_Obj is no longer needed
   (Package_Elaboration) create the set of Why3 keywords at elaboration of the
      package