L927-024 record access needs type entity
The entity of a record component does not contain enough information to
determine in which (sub-)type it lives. Previously, we were using Scope(E) to
determin the record type, but this is incorrect for some subtypes (may point
to the base type). So we extract the information from the prefix of selected
components (in One_Level_Access/Update) and pass it around to the point where
the actual Why3 ident is created (To_Why_Id).

gnat2why-expr.adb:
   (One_Level_Access, One_Level_Update) pass type entity of prefix to
   New_Ada_Record_Access and New_Ada_Record_Update

why-gen-records.adb/.ads:
   (New_Ada_Record_Access, New_Ada_Record_Update) new argument, which is the
   type entity of the record to select from; is passed directly to To_Why_Id

why-inter.adb/.ads:
   (To_Why_Id) new argument Rec, which is only used when the entity is a
   record component. In this case, Rec contains the record (sub-) type which
   should be used for this component.