Add a comment field to Why3 theory nodes, to facilitate debugging
All theory nodes in Why3 files are not preceded by a comment which states why
this theory is created, and from which subprogram in gnat2why. This is meant
to help in reading the generated Why code for debugging.

* gnat2why-decls.adb, gnat2why-driver.adb, gnat2why-expr.adb,
  gnat2why-subprograms.adb, gnat2why-types.adb
Adapt calls to Open_Theory to add a comment argument.
* why-atree-sprint.adb
(Theory_Declaration_Pre_Op): Print comment before theory.
* why-inter.adb
(Open_Theory): Add a comment argument to pass to the builder.
* xtree_sinfo.adb
Add a comment field to the theory declaration node.