[L503-004, LA10-011] Ignore code from instances of standard library generics
When analyzing a unit that is not predefined (that is, this is not the analysis
done for generating Why code for the standard library), ignore bodies defined
in the standard library. In particular, ignore bodies from instances of
generics defined in the standard library. As a result, no warning nor VC is
generated in this case for standard library code.

This avoids generating useless warnings in pedantic mode, on code from the
standard library, since the pedantic mode is meant for projects that use a
different compiler from GNAT. This also avoids generating many VCs on code of
instances of standard library generics, which are of no interest to the user.

* alfa-definition.adb
(Mark_Binary_Op): Do not issue warnings for missing parentheses (in strict
  mode) on code from instances of standard library generics, unless we are
  analyzing the code of a standard library unit.
(Mark_Subprogram_Body): Do not analyze subprogram bodies on code from instances
  of standard library generics, unless we are analyzing the code of a standard
  library unit.

* alfa-util.adb
(Unit_In_Standard_Library): New function returns True on unit from the standard
  library, which includes all units defined in Ada RM, and all units predefined
  by GNAT.
(Location_In_Standard_Library): New function returns True for a location in
  a unit of the standard library.