Following CDR, updates for all "Typo" comments raised by reviewers.
In particular, various conventions have now been adopted and should
be maintained from this point:

"Ada 2012" and "SPARK 2005" have a space in the middle.

All instances of "SPARK95" have been changed to "SPARK 2005".

We now refer to the Ada "RM" NOT the "LRM".

"run-time" is hyphenated when used as an adjective (e.g. "run-time profile")
bu NOT hyphenated when used as a noun (e.g. "blah happens at run time.")

"Precondition" and "Postcondition" are NOT hyphenated, as per the Ada RM.