[LB23-007] More efficient support for Loop_Invariant
The support for Loop_Invariant showed some performance regressions on Astrium
and ACATS tests. It turns out the difference is that, previously, the code
for checking absence of run-time errors in the loop invariant was not
duplicated, while the same for Loop_Invariant used a more straighforward
transformation which duplicated the Why3 code. Use the optimized transformation
now (which is logically equivalent).

* gnat2why-expr-loops.adb
(Wrap_Loop): Modify the transformation of loops.