[LB28-031] Wrong translation of loops with exits and no loop invariant
The translation of loops could be wrong in some cases. This is now corrected.
Additionally, we avoid unrolling the complete loop if there are no loop
invariants in the loop.

* gnat2why-expr-loops.adb
(Wrap_Loop): Make sure the initial statements before the loop, if any, are
  inside the try-block, so that any early exit is captured.
(Get_Loop_Invariant): In the case where there are no loop invariants in the
  loop, put all statements in the Final_Stmts list.