LC19-024 implement range checks based on frontend flags
Use the Do_Range_Check flag to detect whether a range check is needed, do not
systematically insert a range check when going from "int" to the Ada type in
Why.

Unfortunately, Conversion and range checks still are intertwined, because
range checks are always done on "int", so we must insert that check at the
right place in the conversion. This becomes worse in the case of a conversion
between an Ada float and an Ada integer; the conversion then looks like that:

Ada integer -> int -> real -> Ada float

and we have to insert the range check at the right place, depending on whether
the range check is on the integer type or the float type.

gnat2why-expr.adb:
  new function (Get_Range_Check_Info): For a node with the Do_Range_Check
  flag, determine the correct type to do the range check in. This is done by
  looking at the parent node.
  new function (Insert_Range_Check): put a range check on top of the given Why
  expression
  (Transform_Expr): post treatemnt of translated expression is now much
  simpler, because we simply can insert the required checks and/or
  conversions.

why-gen-expr.adb:
  (Insert_Conversion): previously, we could simply add a conversion to force
  a range check. Now we have to find the right point in the conversion chain
  to insert the check. This is done here by expliciting all the cases

why-gen-scalars.adb:
  (Define_Scalar_Conversions): do not generate the program function for the
  "of_int" conversion, instead generate a "range_check" function on the Why
  base type ("int" or "real")

why-gen-arrays.adb:
  (New_Array_Access): No check is generated for the array access itself; this
  is now taken care of by a range check on the index expression.

_gnatprove_standard.mlw
_gnatprove_standard_th.why:
  There are no program functions for arrays anymore. The corresponding
  theories move into the .mlw file, and are renamed.