[LB28-004] Support arbitrary loop variants
Loop variants that occur anywhere in the loop are now translated to Why, in a
conservative way: if a variant is proved, it cannot fail at run time; but an
unprovable VC might be generated for some variants. This is because the
translation relies on the Why loop around the Ada loop invariant, so the
variant should progress between the beginning of this logical loop and its
point of occurrence in the code, and not regress between this point and the end
of the logical loop (to be proved).

* gnat2why-expr-loops.adb
(Transform_Loop_Variant): Adapt to arbitrary loop variants.
(Transform_Loop_Body_Statements): New procedures to translate the initial or
  final part of an Ada loop (before and after the selected (in)variants), doing
  some special handling for variants that occur in these lists of statements.
(Transform_Loop_Statement): Put together the bits corresponding to the unrolled
  part of the loop, and the bits that occur inside the Why loop.
(Wrap_Loop): Change the transformation into Why to account for the bits for
  arbitrary loop variants.

* gnat2why-expr.adb|s
(Transform_Statements): New function to translate a statement list, taking
  into account cutpoints (extracted from the existing Transform_Statements).