[LB23-007] Generate loop invariant VC for non-selected loop invariants
Loop invariants that do not occur in the first batch of loop invariants (and
variants) are said to be "not selected", and treated as assertions. Still, we
prefer a loop invariant VC to be generated, for the user to get a nicer
message (instead of one related to an assertion). Also correct the setting of
reason for loop variant VCs, which was working by luck, as both the Assert and
Loop_Variant reasons were given.

* gnat2why-expr-loops.adb
(Transform_Loop_Body_Statements, Transform_Loop_Statement): Put a single
  reason for loop variant VCs.

* gnat2why-expr.adb
(Transform_Statement): Put the loop invariant reason when needed.

* why-gen-progs.adb|s
(New_Located_Assert): New function takes reason in parameter.