[LC20-029] Use target type of conversion on fixed-point multiplying operands
The multiplication and division operations on fixed-point types have a return
type of universal_fixed (with no bounds), which is used as an overload
resolution trick to allow free conversion between certain real types on the
result of multiplication or division. The target non-universal type determines
the actual sort of multiplication or division performed, and therefore
determines the possibility of overflow. In the compiler, the multiplication is
expanded so the operands are first converted to some common type, so back ends
don't see the universal_fixed Etype. Here, we are seeing the unexpanded
operation because we are running in a mode that disables the expansion. Hence,
we recognize the universal_fixed case specially and in that case use the target
type of the enclosing conversion.

* gnat2why-expr.adb
(Transform_Expr): Special case for multiplication and division on fixed-point
  types.