[M121-012] Improve provability of VCs involving int to real conversions
To get better results when an SMT prover does not natively support the theory
of int to real conversions, instead of generating
  <integral part>.<fractional part>
we generate instead
  Floating.real_of_int (<integral part>) + 0.<fractional part>
so that the SMT prover can reason in the integers for the integral part.

* gnat_backends/why/why-images.adb
(P) on Ureal: Better printing of real literal for SMT provers.